feat: _get_dec_oct_dotted
