+++ src/menu/compute_decimal/to_utils.h
@@ -1,0 +1,7 @@
+#ifndef COMPUTE_DECIMAL_TO_UTILS_H
+#define COMPUTE_DECIMAL_TO_UTILS_H
+#include "./to_utils.c"
+
+extern DottedDecimal _get_dec_bin_dotted(char* dec_input);
+
+#endif
