--- src/menu/compute_decimal/types.h
@@ -1,9 +1,0 @@
-#ifndef COMPUTE_DECIMAL_TYPES_H
-#define COMPUTE_DECIMAL_TYPES_H
-
-typedef struct {
-  char* before_dot;
-  char* after_dot;
-} DottedDecimal;
-
-#endif
